Modified entities and Implemented search and rating endpoints for restaurant

+++ src/main/java/com/example/Foody/Repositories/FoodRepository.java
@@ -1,15 +1,19 @@
 package com.example.Foody.Repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import javax.transaction.Transactional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.example.Foody.Domain.Food;
-import com.example.Foody.Domain.Resturant;
 
+/**
+ * @author Mahesvaran Ashokan
+ *
+ */
 @Repository
 @Transactional
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
 
 	
 }
